requirements.txt
# Web Scraping
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==5.0.0
requests==2.31.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# PDF Processing
pdfplumber==0.10.3
pypdf==3.17.4

# AI Processing
google-generativeai==0.3.2

# Scheduling
apscheduler==3.10.4

# Data Processing
pandas==2.1.4
python-dateutil==2.8.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Frontend
streamlit==1.29.0
plotly==5.18.0
altair==5.2.0

# Utilities
tqdm==4.66.1
retry==0.9.2
